Add dynamic gallery to the "Travel Tips" section
