+++ src/new_achitecture/modules/budgetTravelTips/presentation/BTTStartComponent/GalleryComponent.module.css
@@ -26,6 +26,7 @@
 .image {
     width: 40rem;
     height: 40rem;
+    transition: transform 1s ease;
 }
 
 .navGallery {
@@ -36,21 +37,21 @@
 }
 
 .galBut {
-    border: 2px solid #e7680b;
-    border-radius: 8rem;
-    width: 4rem;
-    height: 4rem;
+    background: none;
+    color: #ff5722;
+    font-size: 2rem;
+    border: none;
+    outline: none;
     cursor: pointer;
-    background-color: white;
-    font-size: 1rem;
-    font-weight: bold;
-    color: #e7680b;
+
+    padding: 0
 }
 
-.galBut:hover {
-    background-color: #ff5722;
-    color: whitesmoke;
-}
+/*.imageContainer {*/
+/*    position: relative;*/
+/*    width: 100%;*/
+/*    overflow: hidden;*/
+/*}*/
 
 
 @media screen and (max-width: 320px) {
@@ -83,14 +84,8 @@
         width: 30rem;
         height: 30rem;
     }
-
-    .galBut {
-        border-radius: 5rem;
-        width: 2.6rem;
-        height: 2.6rem;
-        font-size: 0.5rem;
-    }
 }
+
 @media (max-width: 600px) {
     .navGallery {
         gap: 0.4rem;
@@ -99,13 +94,6 @@
     .image {
         width: 20rem;
         height: 20rem;
-    }
-
-    .galBut {
-        border-radius: 5rem;
-        width: 2.6rem;
-        height: 2.6rem;
-        font-size: 0.5rem;
     }
 }
 
@@ -118,13 +106,6 @@
         width: 17rem;
         height: 17rem;
     }
-
-    .galBut {
-        border-radius: 5rem;
-        width: 2.6rem;
-        height: 2.6rem;
-        font-size: 0.5rem;
-    }
 }
 
 @media (max-width: 380px) {
@@ -136,13 +117,6 @@
         width: 16rem;
         height: 16rem;
     }
-
-    .galBut {
-        border-radius: 4rem;
-        width: 2.2rem;
-        height: 2.2rem;
-        font-size: 0.3rem;
-    }
 }
 
 @media (max-width: 360px) {
@@ -153,12 +127,5 @@
     .image {
         width: 10rem;
         height: 10rem;
-    }
-
-    .galBut {
-        border-radius: 2rem;
-        width: 2.1rem;
-        height: 2.1rem;
-        font-size: 0.3rem;
     }
 }
